Add files via upload
This update adds attacking, and 'looking' at tile descriptions using the L key. the look window is a bit weird, I might fix it later when I have time. When it goes above half way on the screen, it flips upside down. This is intentional, but can be difficult to tell what tile you are looking at.